move framing back to original location
